+++ cmd/gaia/simulation/sim_test.go
@@ -9,7 +9,13 @@ import (
 	"github.com/tendermint/tendermint/libs/log"
 
 	gaia "github.com/cosmos/cosmos-sdk/cmd/gaia/app"
-	// stakesim "github.com/cosmos/cosmos-sdk/x/stake/simulation"
+	"github.com/cosmos/cosmos-sdk/x/mock"
+)
+
+const (
+	NumKeys   = 10
+	NumBlocks = 1000
+	BlockSize = 1000
 )
 
 func TestFullGaiaSimulation(t *testing.T) {
@@ -18,4 +24,16 @@ func TestFullGaiaSimulation(t *testing.T) {
 	db := dbm.NewMemDB()
 	app := gaia.NewGaiaApp(logger, db, nil)
 	require.Equal(t, "GaiaApp", app.Name())
+
+	// Run randomized simulation
+	mock.RandomizedTesting(
+		t, app.BaseApp,
+		[]mock.TestAndRunTx{},
+		[]mock.RandSetup{},
+		[]mock.Invariant{},
+		NumKeys,
+		NumBlocks,
+		BlockSize,
+	)
+
 }
